rename database::comparison to database::metric
