+++ database/src/lib.rs
@@ -9,8 +9,8 @@ use std::ops::{Add, Sub};
 use std::sync::Arc;
 use std::time::Duration;
 
-pub mod comparison;
 pub mod interpolate;
+pub mod metric;
 pub mod pool;
 pub mod selector;
