+++ site/src/api.rs
@@ -253,7 +253,7 @@ pub mod bootstrap {
 pub mod comparison {
     use crate::benchmark_metadata::ProfileMetadata;
     use collector::Bound;
-    use database::{comparison::Metric, Date};
+    use database::{metric::Metric, Date};
     use serde::{Deserialize, Serialize};
     use std::collections::HashMap;
 
@@ -595,7 +595,7 @@ pub mod github {
 
 pub mod triage {
     use collector::Bound;
-    use database::comparison::Metric;
+    use database::metric::Metric;
     use serde::{Deserialize, Serialize};
 
     #[derive(Debug, Clone, Serialize, Deserialize)]
